Fixed compilation error on Grendel
boutmesh.cxx(127): error: more than one instance of overloaded function "sqrt" matches the argument list:
            function "sqrt(double)"
            function "std::sqrt(long double)"
            function "std::sqrt(float)"
            argument types are: (int)
      BoutReal ideal = sqrt(  MX * NPES / ny); // Results in square domains

Added type cast

+++ src/mesh/impls/bout/boutmesh.cxx
@@ -124,7 +124,7 @@ int BoutMesh::load() {
     
     NXPE = -1; // Best option 
     
-    BoutReal ideal = sqrt(MX * NPES / ny); // Results in square domains
+    BoutReal ideal = sqrt(MX * NPES / ((double) ny)); // Results in square domains
 
     for(int i=1; i<= NPES; i++) { // Loop over all possibilities
       //output.write("Testing %d: %d, %d, %d, %d, %d\n",
